refactor: ParentHome 제안받은 돈길 JSX 로직 함수로 분리
